use peers shared panes
