--- src/panes/NotFound.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function NotFound () {
-  return (
-    <div>
-      <h1>What ya doing here? Get out!</h1>
-    </div>
-  )
-}
